// src/app/models/moodelServicePet.ts
export interface ServicePet {
    id?:             any;
    openingDate?:    any;
    closingDate?:    any;
    priority:     string;
    status:       string;
    title:        string;
    comments:     string;
    client:          any;
    employee:        any; 
    nameClient:   string;
    nameEmploye:  string;

}